add integration v8 test
